Pipeline model
We are given a type:   
type pipe = Plus | Minus | Block | Midbar ;;
We want to model a pipeline system, for which we will use a matrix. We will have the following operations: 
-> a) -> print_pipeline = a function that takes in 2 parameters, a pipe matrix and a bool matrix, outputs a matrix with " " where the value in bool matrix on respective index is false, otherwise prints either "+" or "-" or "|" or "x". 
-> b) -> rotate_pipeline = takes a pipe matrix as input and returns a pipe matrix where all "|" transform to "-" and "+","x"
-> c) -> explore_pipeline = takes a pipe matrix as input and returns a bool matrix with values true or false, where true means that the water current reached the spot, and false means it did not. The water springs from the left top corner (0,0). 
The way water flows is as follows:
(rule.1)  If water is in "|" or "+", it can flow to adjacent vertical spots  if there is a "+" or  "|" on an adjacent spot. 
(rule.2)  If water is in "-" or "+", it can flow to adjacent horizontal spots, only if there is a "-" or "+" on an adjacent spot.